+++ src/scientific/readers/all.ts
@@ -1,6 +1,7 @@
 export * from './aperio-svs.ts'
 export * from './cbf.ts'
 export * from './digital-micrograph.ts'
+export * from './tia-ser.ts'
 export * from './envi.ts'
 export * from './fits.ts'
 export * from './gsf.ts'